Share session server actions
Move shared session cookie server actions into a web-next utility so sign-out and account deletion use the same manual Set-Cookie workaround.

https://github.com/hackers-pub/hackerspub/pull/318#discussion_r3440079626
