+++ web-next/src/lib/sessionActions.ts
@@ -1,0 +1,25 @@
+import type { Uuid } from "@hackerspub/models/uuid";
+import { getRequestEvent } from "solid-js/web";
+import {
+  buildExpiredSessionSetCookieHeader,
+  isSecureRequest,
+  readSessionCookie,
+} from "./sessionCookie.ts";
+
+export async function getCurrentSessionId(): Promise<Uuid | null> {
+  "use server";
+  const event = getRequestEvent();
+  return readSessionCookie(event?.request);
+}
+
+export async function removeSessionCookie(): Promise<void> {
+  "use server";
+  const event = getRequestEvent();
+  if (event == null) return;
+  event.response.headers.append(
+    "Set-Cookie",
+    buildExpiredSessionSetCookieHeader({
+      secure: isSecureRequest(event.request),
+    }),
+  );
+}
