+++ web-next/src/routes/(root)/[handle]/settings/account.tsx
@@ -1,7 +1,6 @@
 import { type RouteDefinition, useParams } from "@solidjs/router";
 import { graphql } from "relay-runtime";
 import { createMemo, createSignal, Show } from "solid-js";
-import { getRequestEvent } from "solid-js/web";
 import { createMutation, loadQuery, useRelayEnvironment } from "solid-relay";
 import IconTrash2 from "~icons/lucide/trash-2";
 import { SettingsCardPage } from "~/components/SettingsCardPage.tsx";
@@ -29,10 +28,7 @@ import {
   routePreloadedQuery,
 } from "~/lib/relayPreload.ts";
 import { decodeRouteParam } from "~/lib/routeParam.ts";
-import {
-  buildExpiredSessionSetCookieHeader,
-  isSecureRequest,
-} from "~/lib/sessionCookie.ts";
+import { removeSessionCookie } from "~/lib/sessionActions.ts";
 import type { accountDeleteMutation } from "./__generated__/accountDeleteMutation.graphql.ts";
 import type { accountPageQuery } from "./__generated__/accountPageQuery.graphql.ts";
 
@@ -89,18 +85,6 @@ const accountDeleteMutation = graphql`
     }
   }
 `;
-
-async function removeSessionCookie(): Promise<void> {
-  "use server";
-  const event = getRequestEvent();
-  if (event == null) return;
-  event.response.headers.append(
-    "Set-Cookie",
-    buildExpiredSessionSetCookieHeader({
-      secure: isSecureRequest(event.request),
-    }),
-  );
-}
 
 export default function AccountSettingsPage() {
   const params = useParams();
